LayoutTests/accessibility/mac/malformed-tree-expected.txt

text
treeitem
Entry A


text

text
This tests that a malformed tree is accessible.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS tree.role is 'AXRole: AXOutline'
PASS tree1.role is 'AXRole: AXGroup'
PASS tree2.role is 'AXRole: AXGroup'
PASS itemA != null is true
PASS itemB != null is true
PASS itemC != null is true
PASS itemD != null is true
PASS itemE != null is true
PASS itemF != null is true
PASS successfullyParsed is true

TEST COMPLETE
